requirements.txt
Flask[async]
python-dotenv
litellm
sendgrid
httpx
certifi
Gunicorn